Rename widget warning to warning.js
